src: add tweaks to release post script
- enable the re-use of the tmpdir via another argument, useful when
  generating the release notes for multiple versions - avoids redundant
  git cloning
- support mac's version of `mktemp` and add error checking for failure
  to make a working tmp directory
- update the format for the new website
- output the markdown file directly into the `_posts` directory as
  opposed to printing to stdout

+++ src/gen-release-post.sh
@@ -1,14 +1,25 @@
 #!/bin/sh
 VERSION=$1
 if test -z "$VERSION"; then
-    echo "Usage: $0 VERSION [flux-core|flux-sched]" >&2
+    echo "Usage: $0 VERSION [flux-core|flux-sched] [WORKDIR]" >&2
     exit 1
 fi
 NAME=${2:-flux-core}
-WORKDIR=$(mktemp --directory ${TMPDIR:-/tmp}/${NAME}.XXXXXX)
-echo "Cloning $NAME $VERSION into $WORKDIR" >&2
-git clone --quiet --branch=v${VERSION} https://github.com/flux-framework/${NAME} ${WORKDIR}
-cd ${WORKDIR}
+WORKDIR=${3:-$(mktemp -d ${TMPDIR:-/tmp}/${NAME}.XXXXXX)}
+SCRIPTDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
+
+if [[ -z "${WORKDIR}" ]]; then
+    echo "Failed to make temporary work directory. Exiting"
+    exit 1
+fi
+if [[ ! -d "${WORKDIR}/.git" ]]; then
+    echo "Cloning $NAME $VERSION into $WORKDIR" >&2
+    git clone --quiet --branch=v${VERSION} https://github.com/flux-framework/${NAME} ${WORKDIR}
+    cd ${WORKDIR}
+else
+    cd ${WORKDIR}
+    git checkout v${VERSION}
+fi
 
 URL="https://github.com/flux-framework/${NAME}/issues/"
 TAG=$(git describe)
@@ -16,33 +27,36 @@ AUTHOR="flux-framework"
 COMMIT=$(git rev-parse $TAG)
 DATE=$( git cat-file -p $COMMIT \
       | gawk '/^tagger/{print strftime ("%F %H:%M:%S %z", $(NF-1))}')
+DOWNLOAD="https://github.com/flux-framework/${NAME}/releases/tag/$TAG"
+
+CALDATE=$( git cat-file -p $COMMIT \
+      | gawk '/^tagger/{print strftime ("%F", $(NF-1))}')
+OUTFILE="${SCRIPTDIR}/../_posts/${CALDATE}-$2-$VERSION.md"
+
 
 # Front matter
-cat <<EOF
+cat > ${OUTFILE} <<EOF
 ---
-layout: news_item
-title: "${NAME} v$VERSION Released"
+layout: default
+title: "${NAME} v$VERSION"
 date: "$DATE"
 author: "$AUTHOR"
 categories: release
 version: $VERSION
+download: $DOWNLOAD
 ---
 
-# ${NAME} v$VERSION Released
+Download from GitHub [here]($DOWNLOAD)
 
-<div class="note warning">
-This is an alpha release of ${NAME} and is not intended for production use.
-</div>
-
-Download from GitHub [here](https://github.com/flux-framework/${NAME}/releases/tag/$TAG)
-
-## Release Notes
+# Release Notes
 EOF
 # Get top list of Release Notes for a single version
 #   replace Issue #'s with links
 #
 awk "/^${NAME} version/{n++;getline;getline}; n<=1;" NEWS.md | \
- sed "s|#\([0-9][0-9]*\)|[#\1](${URL}\1)|g"
+ sed "s|#\([0-9][0-9]*\)|[#\1](${URL}\1)|g" >> ${OUTFILE}
 
-cd /tmp 
-rm -rf ${WORKDIR}
+if [[ -z "$3" ]]; then
+    cd /tmp 
+    rm -rf ${WORKDIR}
+fi
